+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.2.91
+
+### Patch Changes
+
+- Updated dependencies [33d92e4]
+    - @sap-ux/odata-service-inquirer@2.14.10
+
+## 0.2.90
+
+### Patch Changes
+
+- Updated dependencies [090cdec]
+    - @sap-ux/odata-service-inquirer@2.14.9
+
+## 0.2.89
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.10.14
+- @sap-ux/abap-deploy-config-inquirer@1.6.79
+- @sap-ux/odata-service-inquirer@2.14.8
+
 ## 0.2.88
 
 ### Patch Changes